LPS-89172 Rename package to util
