+++ modules/apps/portal-template/portal-template-soy-api/src/main/java/com/liferay/portal/template/soy/util/SoyContext.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.template.soy.utils;
+package com.liferay.portal.template.soy.util;
 
 import java.util.Map;
