+++ modules/apps/portal-template/portal-template-soy-api/src/main/java/com/liferay/portal/template/soy/util/SoyRawData.java
@@ -12,17 +12,19 @@
  * details.
  */
 
-package com.liferay.portal.template.soy.utils;
-
-import java.util.Map;
+package com.liferay.portal.template.soy.util;
 
 /**
- * @author Matthew Tambara
+ * This interface can be implemented by any specialized class that is capable
+ * of returning a low level value for the Soy template system. Such low level
+ * value will be passed to Soy engine verbatim, without any smart/additional
+ * type conversion.
+ *
+ * @author Leonardo Barros
+ * @review
  */
-public interface SoyContextFactory {
+public interface SoyRawData {
 
-	public SoyContext createSoyContext();
-
-	public SoyContext createSoyContext(Map<String, Object> context);
+	public Object getValue();
 
 }
