+++ modules/apps/portal-template/portal-template-soy-api/src/main/resources/com/liferay/portal/template/soy/util/packageinfo
